+++ dmr/openapi/views/scalar.py
@@ -33,7 +33,9 @@ class ScalarView(OpenAPIView):
             self.template_name,
             context={
                 'title': self.schema.info.title,
-                'schema': self.dumps(self.schema.convert()),
+                'schema': self.dumps(
+                    self.schema.convert(skip_validation=self.skip_validation),
+                ),
                 'scalar_cdn': cdn_config.get('scalar'),
             },
             content_type=self.content_type,