+++ dmr/openapi/views/stoplight.py
@@ -33,7 +33,9 @@ class StoplightView(OpenAPIView):
             self.template_name,
             context={
                 'title': self.schema.info.title,
-                'schema': self.dumps(self.schema.convert()),
+                'schema': self.dumps(
+                    self.schema.convert(skip_validation=self.skip_validation),
+                ),
                 'stoplight_cdn': cdn_config.get('stoplight'),
             },
             content_type=self.content_type,