+++ dmr/openapi/views/yaml.py
@@ -30,6 +30,8 @@ class OpenAPIYamlView(OpenAPIView):
     def get(self, request: HttpRequest) -> HttpResponse:
         """Render the OpenAPI schema as YAML response."""
         return HttpResponse(
-            content=yaml.safe_dump(self.schema.convert()),
+            content=yaml.safe_dump(
+                self.schema.convert(skip_validation=self.skip_validation),
+            ),
             content_type=self.content_type,
         )