+++ docs/tools/sphinx_ext/run_examples.py
@@ -408,7 +408,10 @@ class _OpenAPIBuilder(_BaseBuilder):
         urlpatterns.append(
             path(
                 self.config['openapi_url'].lstrip('/'),
-                OpenAPIJsonView.as_view(schema),
+                OpenAPIJsonView.as_view(
+                    schema,
+                    skip_validation=self.config.get('skip_validation', False),
+                ),
             ),
         )
         return urlpatterns